V.1
- Node
- Express
- Mongoose
- Models
- Controllers
- Express Middleware
- JWT
- Passive authentication
- Payloads
- Encryption
- Routes
- CORS
- Revalidate records
- MongoDB
- MongoAtlas
- MongoCompass

It's a pretty heavy section in the sense that it's not React that we'll be looking at, but we'll be doing everything with JavaScript, and we should already be fortified with it.

The backend that we will make here is robust and you could use it in real projects, I hope you take advantage of it and above all learn a lot.